fix: some semantics
