--- blocks/sidebar/sidebar.css
@@ -1,7 +1,0 @@
-.sidebar {
-  width: 312px;
-  box-shadow: 0px 0px 4px rgba(231, 232, 235, 0.6), 0px 4px 16px rgba(231, 232, 235, 0.6);
-  border-radius: 16px;
-  padding: 32px 32px 0;
-  box-sizing: border-box;
-}
