+++ profile.html
@@ -73,6 +73,6 @@
     </div>
   </main>
 
-  <script type="module" src="./script/profile.js"></script>
+  <script type="module" src="script/profile.js"></script>
 </body>
 </html>
