+++ script/profile.js
@@ -4,12 +4,12 @@ import {
   userLevel,
   userProf,
   userCountCourses
-} from '../vars';
+} from './vars-profile.js';
 
-import { user } from '../userInfo.js';
+import { user } from './userInfo.js';
 
 userAvatar.src = user.photo;
 userName.textContent = user.name;
 userLevel.textContent = user.level;
 userProf.textContent = user.profession;
-userCountCourses.textContent = `Пройдено ${user.courses} курса`
+userCountCourses.textContent = `Пройдено ${user.courses} курса`;
