+++ script/vars-profile.js
@@ -1,0 +1,5 @@
+export const userAvatar = document.querySelector('.account-overview__photo');
+export const userName = document.querySelector('.account-overview__title');
+export const userLevel = document.querySelector('.account-overview__user-level');
+export const userProf = document.querySelector('.account-overview__user-profession');
+export const userCountCourses = document.querySelector('.account-overview__user-courses');
