+++ script/vars.js
@@ -104,11 +104,3 @@ export const coursesList = [
     statusClass: 'button_type_disabled'
   }
 ]
-
-/* ___ 3. Profile */
-
-export const userAvatar = document.querySelector('.account-overview__photo');
-export const userName = document.querySelector('.account-overview__title');
-export const userLevel = document.querySelector('.account-overview__user-level');
-export const userProf = document.querySelector('.account-overview__user-profession');
-export const userCountCourses = document.querySelector('.account-overview__user-courses');
